[REFACTOR] Migrate JSONFFIEngine to formal namespace
This PR migrates JSONFFIEngine to a formal namespace.
Also list TODOs to further simplify the JSONFFIEngine.
